+++ tests/benchmarks/footprints/src/main.c
@@ -46,7 +46,7 @@ void main(void)
 	};
 
 	k_mem_domain_init(&footprint_mem_domain,
-			  ARRAY_SIZE(mem_parts), mem_parts);
+			  (uint8_t)ARRAY_SIZE(mem_parts), mem_parts);
 #endif /* CONFIG_USERSPACE */
 
 	run_thread_system();
